Include both license files
The license keyword is meant for specifying the license string, not the
file name of the license.  The correct keyword to use for this is
license_files.  Currently the LICENSE.txt file is being included because
license_files defaults to "LICEN[CS]E*", but K5TEST-LICENSE.txt is
missing.  This change switches to the correct keyword, and additionally
includes the K5TEST-LICENSE.txt file.

Signed-off-by: Carl George <carl@george.computer>